Remove Tests & Component 정리
